Bootstrap added, Sign-in/out feature added

+++ socializeUs/base/views.py
@@ -1,11 +1,45 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
 from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login, logout
 
 from .models import Offering, Tag
 from .forms import OfferForm
 # My views
 
+
+def signinPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password is not matching')
+
+    context = {}
+    return render(request, 'base/signup_in.html', context)
+
+def signOut(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -33,19 +67,27 @@ def offerings(request, ofnum):
     return render(request, 'base/offerings.html', context)
 #    return HttpResponse('Offering Page')
 
+@login_required(login_url='login')
 def createOffer(request):
     form = OfferForm()
     if request.method == 'POST':
         form = OfferForm(request.POST)
         if form.is_valid():
-            form.save()
+            offer = form.save(commit=False)
+            offer.providerID = request.user
+            offer.save()
             return redirect('home')
     context = {'form': form}
     return render(request, 'base/create_offering.html', context)
 
+@login_required(login_url='login')
 def updateOffer(request, ofNum):
+
     offer = Offering.objects.get(serviceID=ofNum)
     form = OfferForm(instance=offer)
+
+    if request.user != offer.providerID:
+        return HttpResponse('You are not allowed to update this offer')
 
     if request.method == 'POST':
         form = OfferForm(request.POST, instance=offer)
@@ -55,8 +97,13 @@ def updateOffer(request, ofNum):
     context = {'form':form}
     return render(request, 'base/create_offering.html', context)
 
+@login_required(login_url='login')
 def deleteOffer(request, ofNum):
     offer = Offering.objects.get(serviceID=ofNum)
+    
+    if request.user != offer.providerID:
+        return HttpResponse('You are not allowed to delete this offer')
+
     if request.method == 'POST':
         offer.delete()
         return redirect('home')
